+++ CHANGELOG.rst
@@ -38,6 +38,8 @@ Internal changes:
 - Replace ``black`` and ``flake8`` with ``ruff`` and move configuration of ``pytest`` to ``pyproject.toml``. (:issue:`1007`)
 - Add ``pylint`` for testing code. (:issue:`1014`)
 - Align variable names across the files. (:issue:`1015`)
+- Rework exclusion handling to keep information about excluded coverage. (:issue:`1016`)
+- Add ``mypy`` for testing code. (:issue:`1019`)
 
 8.2 (13 October 2024)
 ---------------------